cursors: add cursors.SVG and replaceColors from cursorimg to enable color replacement in cursor docs
